Moving Candidate Matching code to Candidate Module
